AX: Update accname for presentational images to be ignored despite fallbacks https://bugs.webkit.org/show_bug.cgi?id=289290 rdar://146429365
Reviewed by NOBODY (OOPS!).

Where an image element's alt attribute is set to the empty string (i.e. alt=""), it is intended that the image
is decorative and not conveyed to assistive technology users. Currently, for images
with empty alt that have additional labeling techniques (such as title or aria-label), the image is not ignored
and the additional attributes for labeling are honored.

This commit ensures that decorative images supplied with empty alt are ignored despite the presence
of other labeling methods such as the title attribute.

* LayoutTests/accessibility/img-with-empty-alt-is-ignored-without-fallbacks.html: Added.
* LayoutTests/accessibility/img-no-alt-not-ignored-with-title.html: Deleted.
* Source/WebCore/accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::computeIsIgnored const):